create feature logout and authorization(products, categories & tags)

+++ app/category/controller.js
@@ -1,7 +1,17 @@
 const Category = require('./model');
+const { policyFor } = require('../policy');
 
 async function store(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+
+    if (!policy.can('create', 'Category')) { // <-- can create Category
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat kategori`
+      });
+    }
     let payload = req.body;
     let category = new Category(payload);
     await category.save();
@@ -22,6 +32,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+    if (!policy.can('update', 'Category')) { // <-- can update Category
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate kategori`
+      });
+    }
+
     let payload = req.body;
     let category = await Category.findOneAndUpdate({ _id: req.params.id }, payload, { new: true, runValidators: true });
     return res.json(category);
@@ -41,6 +60,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+    if (!policy.can('delete', 'Category')) { // <-- can delete Category
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus kategori`
+      });
+    }
+
     let deleted = await Category.findOneAndDelete({ _id: req.params.id });
     return res.json(deleted);
   } catch (err) {
